refactor: Rename common to ui_components
