+++ lib/src/components/stim/stim_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:mdigits/src/components/common/centeredbox.dart';
+import 'package:mdigits/src/components/ui_components/centeredbox.dart';
 import 'package:mdigits/src/components/stim/stim_controller.dart';
 
 /// Screen to present stim to participants
