+++ lib/src/components/ui_components/default_appbar.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mdigits/src/components/common/constants.dart';
+import 'package:mdigits/src/components/ui_components/constants.dart';
 
 /// Custom app bar that shows that can show predefined action buttons
 createAppBar({required BuildContext context, bool showActionButtons = false}) {
