+++ lib/src/components/ui_components/loading_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mdigits/src/components/common/default_appbar.dart';
+import 'package:mdigits/src/components/ui_components/default_appbar.dart';
 
 class LoadingView extends StatelessWidget {
   const LoadingView({super.key});
