+++ lib/src/components/ui_components/styles.dart
